Add new files for Work and Shared components
