+++ pages/index.js
@@ -1,10 +1,12 @@
 import Box from '@mui/material/Box';
 import Hero from '../components/Hero';
+import Work from '../components/Work';
 
 export default function Home() {
   return (
     <Box>
       <Hero />
+      <Work />
     </Box>
   )
 }
